requirements.txt
aadhaar==1.0.0
aadhaar-py==1.0.0
absl-py==1.0.0
acct==6.2.0
aiofiles==0.7.0
APScheduler==3.8.1
asgiref==3.4.1
astroid==2.4.2
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
automata-lib==5.0.0
autopep8==1.6.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
black==22.1.0
bs4==0.0.1
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.4
click==8.0.3
colorama==0.4.4
colored==1.4.3
comtypes==1.1.10
configparser==5.0.2
crayons==0.4.0
cryptography==35.0.0
cycler==0.11.0
dateparser==1.1.0
debugpy==1.5.1
decorator==5.1.1
dict-toolbox==2.1.0
distlib==0.3.3
dj-database-url==0.5.0
Django==3.2.9
django-heroku==0.3.1
djangorestframework==3.12.4
dnspython==2.1.0
entrypoints==0.3
executing==0.8.2
fake-useragent==0.1.11
filelock==3.1.0
flake8==2.3.0
Flask==2.0.2
Flask-APScheduler==1.12.2
Flask-Cors==3.0.10
flatbuffers==2.0
fonttools==4.29.0
gast==0.4.0
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.53.0
grains-universal==3
grainsv2==11.10
grpcio==1.42.0
gspread==4.0.1
gunicorn==20.1.0
h11==0.12.0
h5py==3.6.0
httplib2==0.20.2
idem==13.2.1
idem-posix==2.3
idna==2.7
importlib-metadata==4.8.2
ipykernel==6.9.1
ipython==8.0.1
isort==5.6.4
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.1.0
jupyter-client==7.1.2
jupyter-core==4.9.2
kanren==0.2.3
keras==2.7.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
labelImg==1.8.6
lazy-object-proxy==1.4.3
libclang==12.0.0
lxml==4.6.5
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.6.1
mouse==0.7.1
MouseInfo==0.1.3
msgpack==1.0.2
multipledispatch==0.6.0
mypy-extensions==0.4.3
nest-asyncio==1.5.1
nltk==3.7
numpy==1.21.2
oauth2client==4.1.3
oauthlib==3.2.0
opencv-python==4.5.3.56
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.3.4
parso==0.8.3
pathspec==0.9.0
pbr==5.8.0
pep8==1.7.1
pickleshare==0.7.5
Pillow==9.0.1
pipenv==2021.11.23
platformdirs==2.4.0
pop==20.0.0
pop-config==8.0.1
pop-loop==1.0.4
prompt-toolkit==3.0.28
protobuf==3.19.1
proxy-tools==0.1.0
psycopg2==2.9.2
pure-eval==0.2.2
pyaadhaar==1.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycodestyle==2.8.0
pycparser==2.20
pydot==1.4.2
pyflakes==2.1.1
pygame==2.0.1
PyGetWindow==0.0.9
Pygments==2.11.2
pylibjpeg==1.3.0
pylibjpeg-openjpeg==1.1.1
pylibjpeg-rle==1.1.0
pylint==2.6.0
PyMsgBox==1.0.9
pyOpenSSL==21.0.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyperclip==1.8.2
pypiwin32==223
PyQt5==5.14.1
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyRect==0.1.4
PyScreeze==0.1.27
python-dateutil==2.8.2
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.1
pywin32==301
PyYAML==6.0
pyzbar==0.1.8
pyzmq==22.3.0
quora-scraper==1.1.3
regex==2021.10.23
rend==6.3.1
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0.1
scipy==1.7.3
selenium==3.141.0
sip==6.5.1
six==1.15.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.0
SpeechRecognition==3.8.1
sqlparse==0.4.2
stack-data==0.2.0
stevedore==3.5.0
tabulate==0.8.9
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-io-gcs-filesystem==0.22.0
termcolor==1.1.0
threadpoolctl==3.0.0
toml==0.10.1
tomli==2.0.1
toolz==0.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
trio==0.19.0
trio-websocket==0.9.2
tweepy==4.6.0
typing_extensions==4.0.0
tzdata==2021.4
tzlocal==4.0.1
undetected-chromedriver==2.1.1
unification==0.2.2
uritemplate==4.1.1
urllib3==1.25.8
userpaths==0.1.3
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
webdriver-manager==3.5.1
Werkzeug==2.0.2
whitenoise==5.3.0
wikipedia==1.4.0
wrapt==1.12.1
wsproto==1.0.0
xmltodict==0.12.0
yapf==0.29.0
youtube-dl==2021.6.6
zipp==3.6.0
